+++ host/xdc/integration_failover_test.go
@@ -2243,8 +2243,8 @@ func (s *integrationClustersTestSuite) TestLocalNamespaceMigration() {
 		TaskQueue:          sw.DefaultWorkerTaskQueue,
 		WorkflowRunTimeout: time.Second * 30,
 	}, "force-replication", migration.ForceReplicationParams{
-		Namespace:      namespace,
-		RpsPerActivity: 10,
+		Namespace:  namespace,
+		OverallRps: 10,
 	})
 
 	s.NoError(err)
